+++ finagle-mysql/src/main/scala/com/twitter/finagle/mysql/Value.scala
@@ -29,7 +29,7 @@ case object NullValue extends Value
  * bytes.
  *
  * @param typ The MySQL [[Type type]] code for this value.
- * @param charset The [[Charset charset]] encoding of the bytes.
+ * @param charset The [[MysqlCharset charset]] encoding of the bytes.
  * @param isBinary Disambiguates between the text and binary protocol.
  * @param bytes The raw bytes for this value.
  */
@@ -85,7 +85,7 @@ class TimestampValue(val injectionTimeZone: TimeZone, val extractionTimeZone: Ti
     bw.writeByte(cal.get(Calendar.MINUTE))
     bw.writeByte(cal.get(Calendar.SECOND))
     bw.writeIntLE(ts.getNanos / 1000) // sub-second part is written as microseconds
-    RawValue(Type.Timestamp, Charset.Binary, isBinary = true, bytes)
+    RawValue(Type.Timestamp, MysqlCharset.Binary, isBinary = true, bytes)
   }
 
   /**
@@ -124,7 +124,7 @@ object TimestampValue
     value match {
       case raw: RawValue if isTimestamp(raw) =>
         val timestamp = if (!raw.isBinary) {
-          val str = new String(raw.bytes, Charset(raw.charset))
+          val str = new String(raw.bytes, MysqlCharset(raw.charset))
           fromString(str, timeZone)
         } else {
           fromBytes(raw.bytes, timeZone)
@@ -262,19 +262,19 @@ object DateValue extends Injectable[Date] with Extractable[Date] {
     bw.writeShortLE(cal.get(Calendar.YEAR))
     bw.writeByte(cal.get(Calendar.MONTH) + 1) // increment 0 indexed month
     bw.writeByte(cal.get(Calendar.DATE))
-    RawValue(Type.Date, Charset.Binary, true, bytes)
+    RawValue(Type.Date, MysqlCharset.Binary, true, bytes)
   }
 
   /**
    * Value extractor for java.sql.Date
    */
   def unapply(v: Value): Option[Date] = v match {
-    case RawValue(Type.Date, Charset.Binary, false, bytes) =>
-      val str = new String(bytes, Charset(Charset.Binary))
+    case RawValue(Type.Date, MysqlCharset.Binary, false, bytes) =>
+      val str = new String(bytes, MysqlCharset(MysqlCharset.Binary))
       if (str == Zero.toString) Some(Zero)
       else Some(Date.valueOf(str))
 
-    case RawValue(Type.Date, Charset.Binary, true, bytes) =>
+    case RawValue(Type.Date, MysqlCharset.Binary, true, bytes) =>
       Some(fromBytes(bytes))
     case _ => None
   }
@@ -322,13 +322,13 @@ object DateValue extends Injectable[Date] with Extractable[Date] {
 
 object BigDecimalValue extends Injectable[BigDecimal] with Extractable[BigDecimal] {
   def apply(b: BigDecimal): Value = {
-    val str = b.toString.getBytes(Charset(Charset.Binary))
-    RawValue(Type.NewDecimal, Charset.Binary, true, str)
+    val str = b.toString.getBytes(MysqlCharset(MysqlCharset.Binary))
+    RawValue(Type.NewDecimal, MysqlCharset.Binary, true, str)
   }
 
   def unapply(v: Value): Option[BigDecimal] = v match {
-    case RawValue(Type.NewDecimal, Charset.Binary, _, bytes) =>
-      Some(BigDecimal(new String(bytes, Charset(Charset.Binary))))
+    case RawValue(Type.NewDecimal, MysqlCharset.Binary, _, bytes) =>
+      Some(BigDecimal(new String(bytes, MysqlCharset(MysqlCharset.Binary))))
     case _ => None
   }
 }